requirements.txt
# requirements.txt
python-dotenv==1.0.0
gradio==4.19.2
numpy==1.26.4
pandas==2.2.1
Pillow==10.2.0

# Hugging Face ve Model
transformers==4.38.1
torch==2.2.0
accelerate==0.27.2
sentencepiece==0.2.0
protobuf==4.25.2

# RAG ve Vektör Veritabanı
chromadb==0.4.24
sentence-transformers==2.5.1
scikit-learn==1.4.2
faiss-cpu==1.7.4

# Belge İşleme
PyMuPDF==1.23.26
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2

# Yardımcı Araçlar
tqdm==4.66.2
pytest==8.0.2
black==24.2.0
flake8==7.0.0

# API ve İletişim
requests==2.31.0
